+++ modules/tools/compareFeatures/view.js
@@ -16,6 +16,7 @@ define(function (require) {
             "hidden.bs.modal": "setIsActivatedToFalse",
             "click .btn-open-list": "setIsActivatedToTrue",
             "click .btn-infos": "toggleRows",
+            "click .btn-print": "preparePrint",
             "click table .glyphicon-remove": "removeFeatureFromList",
             "change select": function (evt) {
                 this.model.setLayerId(evt.target.value);
@@ -113,7 +114,88 @@ define(function (require) {
                 this.renderErrorModal();
             }
         },
+        preparePrint: function () {
+            var layerModel = Radio.request("ModelList", "getModelByAttributes", {id: this.model.get("layerId")}),
+                themeConfig = Radio.request("Schulinfo", "getThemeConfig"),
+                features = this.model.prepareFeatureListToShow(layerModel.get("gfiAttributes"), themeConfig),
+                rowsToShow = this.$el.find("tr:visible").length,
+                tableBody = this.prepareTableBody(features, rowsToShow),
+                rowWidth = this.calculateRowWidth(tableBody[0], 30),
+                title = "Vergleichsliste - Schulen",
+                pdfDef = {
+                    pageSize: "A4",
+                    pageOrientation: "portrait",
+                    content: [
+                        {
+                            table: {
+                                headerRows: 1,
+                                widths: rowWidth,
+                                body: tableBody
+                            },
+                            layout: {
+                                hLineWidth: function (i, node) {
+                                    return i === 0 || i === node.table.body.length ? 2 : 1;
+                                },
+                                vLineWidth: function (i, node) {
+                                    return i === 0 || i === node.table.widths.length ? 2 : 1;
+                                },
+                                fillColor: function (i) {
+                                    return i % 2 === 0 ? "#eeeeee" : "#dddddd";
+                                }
+                            }
+                        }
+                    ]
+                };
 
+            Radio.trigger("BrowserPrint", "print", "Vergleichsliste_Schulen", pdfDef, title, "download");
+        },
+
+        prepareTableBody: function (features, rowsToShow) {
+            var tableBody = [];
+
+            _.each(features, function (rowFeature, rowIndex) {
+                var row = [];
+
+                if (rowIndex < rowsToShow) {
+                    _.each(rowFeature, function (val) {
+                        if (_.isUndefined(val)) {
+                            row.push("");
+                        }
+                        // header cells get extra styling
+                        else if (rowIndex === 0) {
+                            row.push({
+                                text: String(val),
+                                style: "bold"
+                            });
+                        }
+                        else if (_.isArray(val)) {
+                            row.push(String(val).replace(/,/g, ",\n"));
+                        }
+                        else {
+                            row.push(String(val));
+                        }
+                    });
+                    tableBody.push(row);
+                }
+            });
+            tableBody[0][0] = {
+                text: ""
+            };
+            return tableBody;
+        },
+        calculateRowWidth: function (firstRow, firstRowWidth) {
+            var numDataRows = firstRow.length - 1,
+                rowWidth = [String(firstRowWidth) + "%"],
+                dataRowsWidth = 100 - firstRowWidth,
+                dataRowWidth = String(dataRowsWidth / numDataRows) + "%";
+
+            _.each(firstRow, function (row, index) {
+                if (index > 0) {
+                    rowWidth.push(dataRowWidth);
+                }
+            });
+            return rowWidth;
+        },
         /**
          * shows or hides the rows wiht the class toggle-row
          * and sets the button text